Move MockEntity class into helpers directory
